Removed V notation.  Added progress tracking.
Removed V notation from files because we are now tracking in GitHub.
Did maintain revision notes in the notes in the file.  This version is
Version 1.1

Also added progress tracking so you can see the progress as files are
processed.  This function does not work well if you pipe in a directory
listing because of the way the records are processed.  For best results
use the -PSPath parameter on the function itself to specify the
directory or files of interest.