Simplified Version Check
Originally the #requires version was not in the correct format so I
added code to check major version because I didn't think it worked.
After correcting the form I now use the #requires directive to alert the
user of the minimum required version of Powershell for this function.
Less code = better code